+++ test/fixtures/sort_constructors_first/compact_no_blank_lines.input.dart
@@ -1,0 +1,6 @@
+class Compact {
+  final int a;
+  final int b;
+  void run() {}
+  Compact(this.a, this.b);
+}