+++ test/fixtures/sort_constructors_first/enum_with_constructor.input.dart
@@ -1,0 +1,10 @@
+enum Planet {
+  earth(5.97),
+  mars(0.64);
+
+  final double mass;
+
+  double get relative => mass / 5.97;
+
+  const Planet(this.mass);
+}