+++ test/fixtures/sort_constructors_first/named_and_redirecting.input.dart
@@ -1,0 +1,15 @@
+class Vector {
+  final double x;
+
+  final double y;
+
+  double get magnitude => x + y;
+
+  Vector(this.x, this.y);
+
+  void scale() {}
+
+  Vector.zero() : this(0, 0);
+
+  factory Vector.diagonal(double d) => Vector(d, d);
+}