+++ test/fixtures/sort_constructors_first/static_members_interleaved.input.dart
@@ -1,0 +1,13 @@
+class Counter {
+  static int instances = 0;
+
+  final int start;
+
+  static Counter create() => Counter(0);
+
+  int get current => start;
+
+  Counter(this.start);
+
+  static const int max = 100;
+}